Bug Fix on line 152.
As per the standards followed everywhere in the world, context should directly be passed in apply as first argument but not a property inside an object. Currently, callback is called with the "subscription.object"(line 45 and 48) which has context inside it but the context should directly be sent as first argument and not as a property of an object. eventObject(line 151) should contain its callback and the context in which callback is going to be executed but here the object is passed and in the callback, it is expected to read the context(this) from the passed object.
Fix for the same in this pull request.